+++ src/modules/files/dto/list-files.dto.ts
@@ -31,4 +31,16 @@ export class ListFilesDto {
   @IsOptional()
   @IsString()
   mimeType?: string;
+
+  @IsOptional()
+  @IsString()
+  appId?: string;
+
+  @IsOptional()
+  @IsString()
+  userId?: string;
+
+  @IsOptional()
+  @IsString()
+  purpose?: string;
 }